requirements.txt
flask
scikit-learn
cryptography
pandas
numpy
flask-sqlalchemy
werkzeug
boto3
flask-login
python-dotenv
gunicorn
PyPDF2
